refactored components into subdirectories for organization.
