+++ frontend/src/components/Events/EventCard/EventCard.css
