+++ frontend/src/components/Modals/LoginFormModal/LoginForm.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { useDispatch } from 'react-redux';
 // import { Navigate } from "react-router-dom";
-import * as sessionActions from '../../store/session';
+import * as sessionActions from '../../../store/session';
 
 import './LoginForm.css';
